requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
ecdsa==0.19.1
fastapi==0.115.14
fastapi-contrib==0.2.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
loguru==0.7.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
